+++ src/api/getCompletedPaymentInfo.ts
@@ -1,16 +1,8 @@
-import axios from "axios";
-import { API_BASE_URL } from "@api/config.ts";
+import axiosInstance from "@api/axiosInstance";
 
-export interface PaymentInfoProps {
-  headers: { [key: string]: string };
-  count: number;
-}
-
-export const getCompletedPaymentInfo = async ({
-  count,
-  headers
-}: PaymentInfoProps) => {
-  const API_URL = `${API_BASE_URL}/carts/orders/payments?count=${count}`;
-  const response = await axios.get(API_URL, { headers });
+export const getCompletedPaymentInfo = async (count: number) => {
+  const response = await axiosInstance.get(
+    `/carts/orders/payments?count=${count}`
+  );
   return response.data.data;
 };
